test: ✅ extract CashRemunerationTransactions test fixtures
The spec exceeded the 300-line max-lines rule. Move stubs, query-result
builders, table prop accessors, address constants and createWrapper into
a *.fixture.ts. vi.mock/vi.hoisted stay in the spec — vitest cannot
export hoisted bindings, so mocks remain local per the repo's fixture
pattern.

Refs #2036
